refactor: address form in separate file; fix: NavLink instead of Link
