+++ src/components/RegistrationForm/RegistrationForm.tsx
@@ -5,31 +5,29 @@ import {
   Grid,
   Typography,
   Box,
-  Link,
   FormHelperText,
   FormControlLabel,
   Checkbox,
 } from "@mui/material";
 import { useState } from "react";
-import type { Dispatch, SetStateAction } from "react";
-import { Link as RouterLink } from "react-router";
+import { NavLink } from "react-router";
 import { DatePicker } from "@mui/x-date-pickers/DatePicker";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers";
 import type {
   UserRegistration,
   UserRegistrationErrorText,
-  AddressErrorText,
 } from "../../shared/types/UserRegistration";
-import type { Address } from "../../shared/types/Address";
 import "./RegistrationForm.css";
 import dayjs from "dayjs";
+import AddressForm from "./AddressForm";
 
 type RegistrationFormProps = {
   userProps?: UserRegistration;
   errorTextProps?: UserRegistrationErrorText;
   onUserPropsChange?: () => void;
   onSubmit?: () => void;
+  submitError?: string | null;
 };
 
 function RegistrationForm({
@@ -37,6 +35,7 @@ function RegistrationForm({
   errorTextProps,
   onUserPropsChange,
   onSubmit,
+  submitError,
 }: RegistrationFormProps): React.ReactElement {
   const [showPassword, setShowPassword] = useState(false);
   const [useOneAddress, setUseOneAddress] = useState(false);
@@ -161,12 +160,12 @@ function RegistrationForm({
               }
               label="Use one address for both shipping and billing"
             />
-            <AddressField
+            <AddressForm
               addressTitle="Shipping address"
               address={shippingAddress}
               setAddress={setShippingAddress}
             />
-            <AddressField
+            <AddressForm
               addressTitle="Billing address"
               address={billingAddress}
               setAddress={setBillingAddress}
@@ -198,6 +197,11 @@ function RegistrationForm({
                 </FormHelperText>
               )}
             </Grid>
+            {!!submitError && (
+              <FormHelperText error sx={{ mx: 0 }}>
+                {submitError}
+              </FormHelperText>
+            )}
             <Grid>
               <Button
                 fullWidth
@@ -213,143 +217,15 @@ function RegistrationForm({
             </Grid>
             <Grid container>
               <Grid>
-                <Link component={RouterLink} to="/login" variant="body2">
+                <NavLink to="/login">
                   {"Have an account already? Sign In"}
-                </Link>
+                </NavLink>
               </Grid>
             </Grid>
           </Grid>
         </form>
       </Box>
     </Container>
-  );
-}
-
-type AddressFormProps = {
-  addressTitle: string;
-  address: Address;
-  setAddress: Dispatch<SetStateAction<Address>>;
-  errorText?: AddressErrorText;
-  useDefaultAddress?: boolean;
-  setUseDefaultAddress?: Dispatch<SetStateAction<boolean>>;
-  isDisabled?: boolean;
-};
-
-function AddressField({
-  addressTitle,
-  address,
-  setAddress,
-  errorText,
-  useDefaultAddress,
-  setUseDefaultAddress,
-  isDisabled,
-}: AddressFormProps): React.ReactElement {
-  return (
-    <Box>
-      <Box
-        display={"grid"}
-        alignItems={"start"}
-        gap={1}
-        sx={{ gridTemplateColumns: { xs: "1fr", sm: "1fr 1fr" }, alignContent: "start" }}
-      >
-        <Typography variant="h5" component="h1" align="left" gutterBottom my={0.5}>
-          {addressTitle}
-        </Typography>
-        <FormControlLabel
-          control={
-            <Checkbox
-              checked={useDefaultAddress}
-              onChange={() => setUseDefaultAddress}
-              color="primary"
-            />
-          }
-          label="Use address as default"
-        />
-      </Box>
-      <Box
-        display={"grid"}
-        alignItems={"center"}
-        gap={1}
-        sx={{ gridTemplateColumns: { xs: "1fr", sm: "1fr 1fr" }, alignContent: "center" }}
-      >
-        <Grid>
-          <TextField
-            fullWidth
-            label="Country*"
-            name="country"
-            type="text"
-            variant="outlined"
-            disabled={isDisabled}
-            value={address.country}
-            onChange={() => setAddress}
-          />
-          {!!errorText?.countryErrorText && (
-            <FormHelperText error sx={{ mx: 0 }}>
-              {errorText?.countryErrorText}
-            </FormHelperText>
-          )}
-        </Grid>
-        <Grid>
-          <TextField
-            fullWidth
-            label="City*"
-            name="city"
-            type="text"
-            variant="outlined"
-            disabled={isDisabled}
-            value={address.city}
-            onChange={() => setAddress}
-          />
-          {!!errorText?.cityErrorText && (
-            <FormHelperText error sx={{ mx: 0 }}>
-              {errorText?.cityErrorText}
-            </FormHelperText>
-          )}
-        </Grid>
-      </Box>
-      <Box
-        display={"grid"}
-        alignItems={"center"}
-        gap={1}
-        mt={1}
-        sx={{ gridTemplateColumns: { xs: "1fr", sm: "3fr 2fr" }, alignContent: "center" }}
-      >
-        <Grid>
-          <TextField
-            fullWidth
-            label="Street*"
-            name="street"
-            type="text"
-            variant="outlined"
-            disabled={isDisabled}
-            value={address.street}
-            onChange={() => setAddress}
-          />
-          {!!errorText?.streetErrorText && (
-            <FormHelperText error sx={{ mx: 0 }}>
-              {errorText?.streetErrorText}
-            </FormHelperText>
-          )}
-        </Grid>
-        <Grid>
-          <TextField
-            fullWidth
-            label="Postal Code*"
-            name="postalCode"
-            type="text"
-            variant="outlined"
-            disabled={isDisabled}
-            value={address.postalCode}
-            onChange={() => setAddress}
-          />
-          {!!errorText?.postalCodeErrorText && (
-            <FormHelperText error sx={{ mx: 0 }}>
-              {errorText?.postalCodeErrorText}
-            </FormHelperText>
-          )}
-        </Grid>
-      </Box>
-    </Box>
   );
 }
